[REFACTOR] : LottoPrize와 LottoManager로 class 병합
