+++ __tests__/LottoManager.test.js
@@ -1,3 +1,4 @@
+import Lotto from "../src/domain/Lotto";
 import LottoManager from "../src/domain/LottoManager";
 
 test("구입 금액에 해당하는 만큼 로또를 발행해야 한다", () => {
